observability-service/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.35
alembic>=1.14.0
psycopg2-binary>=2.9.5

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Authentication
pyjwt>=2.8.0,<3.0.0
cryptography>=41.0.0

# HTTP client
httpx>=0.28.0

# Redis
redis>=5.2.0

# Logging
loguru>=0.7.0
structlog>=24.0.0

# LangChain (agentic)
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0

# Observability clients
prometheus-api-client>=0.5.0
elasticsearch>=8.0.0
kubernetes>=28.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
